11/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
sqlalchemy>=2.0.36
aiosqlite==0.19.0
jinja2==3.1.3
python-dotenv==1.0.1
httpx==0.27.0
stripe==7.11.0 